added friends the table itself

+++ application/views/upload_form.php
@@ -26,25 +26,27 @@ Private:<input type="radio" name="type" value="Private" <?php echo set_radio('ty
 </form>
 <h2> Current Design </h2>
 <table border = '1'> 
-<tr><th>Image_Id</th><th>Title</th><th>Price($)</th><th>Type</th><th>Image_Path</th></tr>
+<tr><th>Title</th><th>Image_Path</th><th>Price($)</th><th>Type</th><th>Design_ID</th><th>Action</th></tr>
 <?php
     $customer_id = 1;
     $this->db->where('customer_id', 1); 
     $query = $this->db->get('design');
-    echo "Total Record Returned: " ;echo $query->num_rows();
-    
+    echo "Total Record Returned: " .  $query->num_rows();
     if ($query->num_rows() > 0)
     {
     foreach ($query->result() as $row)
     {
+       if($row->type != 'remove'){
        echo "<tr>";
-       echo "<td>". $row->design_id . "</td>";
        echo "<td>". $row->title . "</td>";
+       echo "<td><img src='" . $row->image_path . "' width=42' height='42'/></td>";
        echo "<td>" . $row->price . "</td>";
        echo "<td>" . $row->type . "</td>";
-       echo "<td><img src='" . $row->image_path . "' width=50px' height='50px'/></td>";
+       echo "<td>". $row->design_id . "</td>";
+       $design_id = $row->design_id;
+       echo "<td><a href=" . site_url("upload/remove?id=". $design_id) ." onclick=\"return confirm('Are you sure you want to delete the remove this design')\">[Remove]</a></td>";
        echo "</tr>";
-   
+       }
     }
     }
 ?>
